fix(hmr): location.reload errors in service worker
add new vite client modification so location.reload falls back to chrome.runtime.reload
